后台验证码 登陆界面修改

+++ myproject-console/src/main/java/com/myproject/console/controller/AdminController.java
@@ -11,12 +11,11 @@ import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.util.CollectionUtils;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 
@@ -44,7 +43,6 @@ public class AdminController {
 
     /**
      * 列表页展示
-     *
      * @param
      * @param model
      * @return
@@ -66,29 +64,46 @@ public class AdminController {
      * @return
      */
     @GetMapping("edit")
-    public String edit(Long id, Model model) {
+    public String edit(Long id, Model model, @RequestParam(name = "pageNum", defaultValue = "1") Integer pageNum,
+                       @RequestParam(name = "pageSize", defaultValue = "10") Integer pageSize) {
+        model.addAttribute("pageNum", pageNum);
+        model.addAttribute("pageSize", pageSize);
         model.addAttribute("admin", adminService.selectByKey(id));
         return "admin/edit";
     }
 
+    /**
+     * 添加页展示
+     */
+    @GetMapping("add")
+    public String add(Model model, @RequestParam(defaultValue = "1") Integer pageNum, @RequestParam(defaultValue =
+            "10") Integer pageSize) {
+        model.addAttribute("pageNum", pageNum);
+        model.addAttribute("pageSize", pageSize);
+        return "admin/add";
+    }
 
     /**
      * 提交保存
-     *
      * @param
      * @return
      */
     @PostMapping("submit")
-    public String submit(Admin admin, RedirectAttributes attr) {
+    public String submit(Admin admin, RedirectAttributes attr, @RequestParam(name = "pageNum", defaultValue = "1")
+            Integer pageNum, @RequestParam(name = "pageSize", defaultValue = "10") Integer pageSize) {
         //判断Admin是否存在
+        String url = "redirect:list";
         if (admin.getId() != null) {
             //编辑
             if (StringUtils.isNotBlank(admin.getPassword())) {
                 admin.setPassword(DigestUtils.md5Hex(admin.getPassword()));
             }
+            url = "redirect:list?pageNum=" + pageNum + "&pageSize=" + pageSize;
         } else {
             //新增
             admin.setPassword(DigestUtils.md5Hex(admin.getPassword()));
+            url = "redirect:list?pageSize=" + pageSize;
+
         }
         Admin a;
         if (StringUtils.isNotBlank(admin.getPassword())) {
@@ -97,18 +112,20 @@ public class AdminController {
             a = adminService.saveOrUpdateIgnore(admin, "password");
         }
         attr.addFlashAttribute("msg", "操作成功");
-        return "redirect:list";
+        return url;
     }
 
     /**
      * 删除
-     *
      * @param ids
      * @return
      */
     @PostMapping("delete")
     @ResponseBody
     public Result<String> delete(Long[] ids) {
+        if (CollectionUtils.isEmpty(Arrays.asList(ids))) {
+            return Result.exception(null, "请选择删除对象");
+        }
         adminService.delete(ids);
         return Result.success(null, "操作成功");
     }
